+++ js/src/cli.ts
@@ -123,6 +123,10 @@ const argv = yargs(hideBin(process.argv))
       "How one query is being processed for multiple accounts: in parallel (true) or sequentially (false). By default - in parallel",
     default: true,
   })
+  .option("parallel-threshold", {
+    type: "number",
+    description: "The maximum number of parallel queries"
+  })
   .option("csv.destination-folder", {
     type: "string",
     alias: "csv.destination",
@@ -421,6 +425,7 @@ async function main() {
   let options: AdsQueryExecutorOptions = {
     skipConstants: argv.skipConstants,
     parallelAccounts: argv.parallelAccounts,
+    parallelThreshold: argv.parallelThreshold,
     dumpQuery: argv.dumpQuery,
   };
   logger.info(`Found ${scriptPaths.length} script to process`);